# services/gateway/src/openapi.yaml
openapi: 3.0.3
info:
  title: Dental Clinic Appointment API
  version: "1.0.0"
  description: >
    Public API surface for the dental appointment booking system, served
    behind the gateway. Internally this fans out to independent
    microservices (patient, dentist, appointment, reminder, notification,
    report) — but every route below is reached the same way it always was,
    as a single relative `/api/*` origin.
servers:
  - url: /api
    description: Through the gateway (relative — works from any host it's deployed behind)
  - url: http://localhost:4000/api
    description: Local dev
...
paths:
...
  /reminders:
    get:
      tags: [Reminders]
      summary: Reminder log
      parameters:
        - name: appointmentId
          in: query
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Reminder" }
  /reminders/{id}/send:
    post:
      tags: [Reminders]
      summary: Send a reminder immediately (demo helper — doesn't wait for the scheduled time)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Sent
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SendResult" }
        "400":
          description: Send failed
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SendResult" }
  /reports/summary:
    get:
      tags: [Reports]
      summary: Clinic-wide reporting metrics
      description: >
        Composed live from every other service (no local database) — fails
        with 503 if any of them is unreachable, since a partial dashboard
        number would be worse than a visible error.
      parameters:
        - name: from
          in: query
          schema: { type: string, format: date-time }
        - name: to
          in: query
          schema: { type: string, format: date-time }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ReportSummary" }
        "503":
          description: report data temporarily unavailable
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /reports/summary/pdf:
    get:
      tags: [Reports]
      summary: Same report as /reports/summary, as a downloadable PDF
      description: >
        Composes the same data as GET /reports/summary and renders it as a
        formatted PDF (clinic header, summary stats, status breakdown,
        per-dentist load) instead of JSON — used by the frontend dashboard's
        "Download PDF" button.
      parameters:
        - name: from
          in: query
          schema: { type: string, format: date-time }
        - name: to
          in: query
          schema: { type: string, format: date-time }
      responses:
        "200":
          description: PDF file
          content:
            application/pdf:
              schema: { type: string, format: binary }
        "503":
          description: report data temporarily unavailable
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
components:
...
  schemas:
    Error:
      type: object
      properties:
        error: { type: string }
...
    Reminder:
      type: object
      properties:
        id: { type: integer }
        appointment_id: { type: integer }
        channel: { type: string, enum: [email, sms] }
        offset_hours: { type: number }
        scheduled_for: { type: string, format: date-time }
        status: { type: string, enum: [pending, sending, sent, failed, skipped] }
        sent_at: { type: string, format: date-time, nullable: true }
        detail: { type: string, nullable: true, description: Preview URL or error message }
        patient_email: { type: string }
        patient_name: { type: string }
        dentist_name: { type: string, nullable: true }
        appt_start_time: { type: string, format: date-time }
        appt_reason: { type: string, nullable: true }
        created_at: { type: string, format: date-time }
    SendResult:
      type: object
      properties:
        ok: { type: boolean }
        detail: { type: string, description: Preview URL (test mode) or SMTP message id / error }
    ReportSummary:
      type: object
      properties:
        total: { type: integer }
        counts:
          type: object
          properties:
            booked: { type: integer }
            completed: { type: integer }
            cancelled: { type: integer }
            no_show: { type: integer }
        noShowRate: { type: number, description: Percentage of finished appointments }
        cancellationRate: { type: number, description: Percentage of finished appointments }
        upcoming: { type: integer }
        remindersSent: { type: integer }
        patients: { type: integer }
        perDentist:
          type: array
          items:
            type: object
            properties:
              id: { type: integer }
              name: { type: string }
              kept: { type: integer }
              no_shows: { type: integer }
              cancelled: { type: integer }
        today: { type: integer, description: Appointments starting today }
        thisMonth: { type: integer, description: Appointments starting in the current calendar month }
        dailySeries:
          type: array
          description: Zero-filled daily appointment counts for the last 14 days
          items:
            type: object
            properties:
              date: { type: string, format: date }
              count: { type: integer }